Audit follow-ups: Tax Invoice overlay, distribution tiers, decimal stock, near-dup catalog warn
#1 Tax Invoice print now applies the InvoiceItemAdjustment overlay
   before grouping. The document the buyer reconciles against FBR
   Annexure-A must mirror what was filed; the customer-facing Bill
   print continues to use real bill-row values.

#2 StockMovement.Quantity + OpeningStockBalance.Quantity promoted
   from int to decimal(18,4) — matches InvoiceItem / DeliveryItem /
   PurchaseItem precision. Pre-fix, sales of fractional UOMs (KG,
   Liter, Carat) truncated at the StockService boundary and on-hand
   drifted upward. IStockService.RecordMovementAsync / GetOnHand{,Bulk}
   signatures + StockRequirement / StockShortage records updated to
   match. New migration PromoteStockQuantityToDecimal alters both
   columns and drops the legacy InvoiceItemId1 shadow column.

#3 Frontend multi-row distribution now mirrors the backend's
   three-tier preference (in-band → above-ceiling → below-floor).
   Pre-fix the per-row picker only knew "in band" vs "any closest";
   it could pick a below-floor divisor when an above-ceiling one was
   the safer audit direction.

#4 Snapshot stale hint now trips on bank-side drift too (band edges
   or weighted-avg cost moving) in addition to numerical drift. The
   hint text differentiates "bill state changed" from "purchase bank
   changed — likely an Annexure-A import".

#5 EditBillForm load fans out invoice / units / item-types in
   parallel without companyId for fast first paint, then re-fires
   item-types with companyId after invoice resolves so the per-company
   on-hand sort + chips appear on top of the legacy-sorted list.

#7 Multi-row distribution skips lineTotal=0 rows instead of
   emitting qty=1 × Rs.0 nonsense.

#10 STOCKMOVEMENT_OVERLAY_SYNC_V1 backfill marker bumps to Warning
    (HTTP 207) when per-bill failures occurred, with the failed bill
    ids inlined in the message — surfaces in standard audit-log
    warning filters instead of being buried under per-bill WARN rows.

#11 ItemTypeService.CreateAsync blocks near-duplicate names — exact
    case-folded match, singular/plural collapse, and one-character
    edit distance for names ≥ 4 chars. Operators can override by
    marking the new row Favorite. Prevents the "Hardware Item" vs
    "Hardware Items" landmine that silently broke optimization for
    bill 3762.

+++ Controllers/StockController.cs
@@ -89,9 +89,11 @@ namespace MyApp.Api.Controllers
             {
                 var it = itemTypes.GetValueOrDefault(id);
                 if (it == null) continue;
-                int opening = openings.GetValueOrDefault(id);
-                int totalIn = moveAggs.Where(x => x.ItemTypeId == id && x.Direction == StockMovementDirection.In).Sum(x => x.Qty);
-                int totalOut = moveAggs.Where(x => x.ItemTypeId == id && x.Direction == StockMovementDirection.Out).Sum(x => x.Qty);
+                // 2026-05-12: decimal opening + totalIn/Out matches the
+                // promoted StockMovement / OpeningStockBalance columns.
+                decimal opening = openings.GetValueOrDefault(id);
+                decimal totalIn = moveAggs.Where(x => x.ItemTypeId == id && x.Direction == StockMovementDirection.In).Sum(x => x.Qty);
+                decimal totalOut = moveAggs.Where(x => x.ItemTypeId == id && x.Direction == StockMovementDirection.Out).Sum(x => x.Qty);
                 rows.Add(new StockOnHandRowDto
                 {
                     ItemTypeId = id,
